// src/types/project-document-type.ts
import { Document, Types } from 'mongoose';
export type ProjectType = Document & {
  projectCode: string;
  name: string;
  budget: number;
  advance: number;
  due: number;
  totalPaid: number;
  startDate: string;
  endDate: string;
  demoLink: string;
  typeOfWeb: string;
  description: string;
  status: boolean;
  clients: Types.ObjectId[];
  projectManager: Types.ObjectId;
  paymentList: Types.ObjectId[];
  reCalculateAll: () => Promise<void>;
  createdAt: Date;
  updatedAt: Date;
};